cmvn and db norm
